requirements.txt
loguru
tqdm
pandas-datareader
yfinance
scikit-learn
plotly